# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NodeWatch API",
        "description": "Self-hosted uptime monitoring service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/monitor": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "List monitors (paginated)",
                "operationId": "6ab0b2421e6d268bcc3ac6648905d15c",
                "responses": {
                    "200": {
                        "description": "Paginated list of monitors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MonitorResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Create a monitor",
                "operationId": "6db749fac833ffa866c3d4cc5879334d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "url",
                                    "method",
                                    "check_interval",
                                    "timeout",
                                    "expected_status",
                                    "regions",
                                    "is_active"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My Website",
                                        "minLength": 3
                                    },
                                    "url": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://example.com"
                                    },
                                    "method": {
                                        "type": "string",
                                        "example": "HEAD",
                                        "enum": [
                                            "GET",
                                            "POST",
                                            "HEAD"
                                        ]
                                    },
                                    "check_interval": {
                                        "type": "integer",
                                        "example": 60,
                                        "enum": [
                                            30,
                                            60,
                                            120,
                                            180,
                                            240,
                                            360
                                        ]
                                    },
                                    "timeout": {
                                        "type": "integer",
                                        "example": 30,
                                        "maximum": 60,
                                        "minimum": 5
                                    },
                                    "expected_status": {
                                        "type": "integer",
                                        "example": 200,
                                        "minimum": 200
                                    },
                                    "regions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "eu-west",
                                                "us-east",
                                                "ap-south"
                                            ]
                                        },
                                        "example": [
                                            "eu-west"
                                        ]
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Monitor created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MonitorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/monitor/{id}": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get a monitor",
                "operationId": "860335bfb964085a74aa924cda04d2d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monitor details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MonitorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Update a monitor (partial)",
                "operationId": "7a80eb67d5f69fc37979fc4466022778",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My Website",
                                        "minLength": 3
                                    },
                                    "method": {
                                        "type": "string",
                                        "example": "HEAD",
                                        "enum": [
                                            "GET",
                                            "POST",
                                            "HEAD"
                                        ]
                                    },
                                    "check_interval": {
                                        "type": "integer",
                                        "example": 60,
                                        "enum": [
                                            30,
                                            60,
                                            120,
                                            180,
                                            240,
                                            360
                                        ]
                                    },
                                    "timeout": {
                                        "type": "integer",
                                        "example": 30,
                                        "maximum": 60,
                                        "minimum": 5
                                    },
                                    "expected_status": {
                                        "type": "integer",
                                        "example": 200,
                                        "minimum": 200
                                    },
                                    "regions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "eu-west",
                                                "us-east",
                                                "ap-south"
                                            ]
                                        }
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Monitor updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MonitorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Delete a monitor",
                "operationId": "b3cb0b88daecac9b50e59a7f4da03cc5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monitor deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MonitorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MonitorResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "type": {
                        "type": "string",
                        "example": "monitors"
                    },
                    "attributes": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "My Website"
                            },
                            "url": {
                                "type": "string",
                                "format": "uri",
                                "example": "https://example.com"
                            },
                            "method": {
                                "type": "string",
                                "example": "HEAD",
                                "enum": [
                                    "GET",
                                    "POST",
                                    "HEAD"
                                ]
                            },
                            "check_interval": {
                                "type": "integer",
                                "example": 60
                            },
                            "timeout": {
                                "type": "integer",
                                "example": 30
                            },
                            "expected_status": {
                                "type": "integer",
                                "example": 200
                            },
                            "regions": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "eu-west"
                                }
                            },
                            "is_active": {
                                "type": "boolean",
                                "example": true
                            },
                            "last_status": {
                                "type": "integer",
                                "example": 0,
                                "enum": [
                                    0,
                                    1,
                                    2
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "relationships": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}